Agents - unify the view mode and the view service
