+++ src/vs/sessions/contrib/changes/browser/changesView.ts
@@ -73,14 +73,13 @@ import { REVEAL_CI_CHECKS_COMMAND_ID } from './checksActions.js';
 import { AGENT_HOST_SKILL_BUTTON_UPDATE_PR_ID, isAgentHostSkillButtonId } from '../../providers/agentHost/browser/agentHostSkillButtons.js';
 import { ActiveSessionContextKeys, CHANGES_VIEW_CONTAINER_ID, CHANGES_VIEW_ID, ChangesContextKeys, ChangesViewMode, IsolationMode, SESSIONS_CHANGES_OPEN_SINGLE_FILE_DIFF_SETTING } from '../common/changes.js';
 import { buildTreeChildren, ChangesTreeElement, ChangesTreeRenderer, IChangesFileItem, IChangesTreeRootInfo, isChangesFileItem, toIChangesFileItem } from './changesViewRenderer.js';
-import { ChangesViewModel } from './changesViewModel.js';
-import { IChangesViewService } from './changesViewService.js';
 import { ResourceTree } from '../../../../base/common/resourceTree.js';
 import { structuralEquals } from '../../../../base/common/equals.js';
 import { compareFileNames, comparePaths } from '../../../../base/common/comparers.js';
 import { IViewsService } from '../../../../workbench/services/views/common/viewsService.js';
 import { ServiceCollection } from '../../../../platform/instantiation/common/serviceCollection.js';
 import { IMarkdownString } from '../../../../base/common/htmlContent.js';
+import { IChangesViewService } from '../common/changesViewService.js';
 
 const $ = dom.$;
 
@@ -93,9 +92,9 @@ const RUN_SESSION_CODE_REVIEW_ACTION_ID = 'sessions.codeReview.run';
 class ChangesMenuWorkbenchButtonBarWidget extends Disposable {
 	constructor(
 		container: HTMLElement,
-		viewModel: ChangesViewModel,
 		hasGitOperationInProgressObs: IObservable<boolean>,
 		@IMenuService menuService: IMenuService,
+		@IChangesViewService changesViewService: IChangesViewService,
 		@IContextKeyService contextKeyService: IContextKeyService,
 		@IContextMenuService contextMenuService: IContextMenuService,
 		@IKeybindingService keybindingService: IKeybindingService,
@@ -105,7 +104,7 @@ class ChangesMenuWorkbenchButtonBarWidget extends Disposable {
 		super();
 
 		const outgoingChangesObs = derivedObservableWithCache<number | undefined>(this, (reader, lastValue) => {
-			const activeSessionState = viewModel.activeSessionStateObs.read(reader);
+			const activeSessionState = changesViewService.activeSessionStateObs.read(reader);
 			const hasGitOperationInProgress = hasGitOperationInProgressObs.read(reader);
 			if (hasGitOperationInProgress) {
 				return lastValue;
@@ -125,7 +124,7 @@ class ChangesMenuWorkbenchButtonBarWidget extends Disposable {
 
 		this._register(autorun(reader => {
 			const hasGitOperationInProgress = hasGitOperationInProgressObs.read(reader);
-			const sessionResource = viewModel.activeSessionResourceObs.read(reader);
+			const sessionResource = changesViewService.activeSessionResourceObs.read(reader);
 			const outgoingChanges = outgoingChangesObs.read(reader) ?? 0;
 
 			const buttonBar = new MenuWorkbenchButtonBar(
@@ -229,8 +228,8 @@ class ChangesMenuWorkbenchButtonBarWidget extends Disposable {
 class ChangesWorkbenchButtonBarWidget extends Disposable {
 	constructor(
 		container: HTMLElement,
-		viewModel: ChangesViewModel,
 		@IMenuService menuService: IMenuService,
+		@IChangesViewService changesViewService: IChangesViewService,
 		@IContextKeyService contextKeyService: IContextKeyService,
 		@IInstantiationService instantiationService: IInstantiationService,
 	) {
@@ -254,12 +253,12 @@ class ChangesWorkbenchButtonBarWidget extends Disposable {
 		});
 
 		const operationActionGroupsObs = derived<IAction[][]>(reader => {
-			const changeset = viewModel.activeSessionChangesetObs.read(reader);
+			const changeset = changesViewService.activeSessionChangesetObs.read(reader);
 			if (!changeset) {
 				return [];
 			}
 
-			const operations = viewModel.activeSessionChangesetOperationsObs.read(reader);
+			const operations = changesViewService.activeSessionChangesetOperationsObs.read(reader);
 			const changesetOperations = operations
 				.filter(op => op.scopes.includes(SessionChangesetOperationScope.Changeset));
 
@@ -384,8 +383,6 @@ export class ChangesViewPane extends ViewPane {
 	private currentBodyHeight = 0;
 	private currentBodyWidth = 0;
 
-	readonly viewModel: ChangesViewModel;
-
 	constructor(
 		options: IViewPaneOptions,
 		@IKeybindingService keybindingService: IKeybindingService,
@@ -397,16 +394,14 @@ export class ChangesViewPane extends ViewPane {
 		@IOpenerService openerService: IOpenerService,
 		@IThemeService themeService: IThemeService,
 		@IHoverService hoverService: IHoverService,
+		@IChangesViewService private readonly changesViewService: IChangesViewService,
 		@IEditorService private readonly editorService: IEditorService,
 		@ISessionsService private readonly sessionsService: ISessionsService,
 		@ILabelService private readonly labelService: ILabelService,
 		@ILogService private readonly logService: ILogService,
 		@ITelemetryService private readonly telemetryService: ITelemetryService,
-		@IChangesViewService changesViewService: IChangesViewService,
 	) {
 		super({ ...options, titleMenuId: MenuId.ChatEditingSessionTitleToolbar }, keybindingService, contextMenuService, configurationService, contextKeyService, viewDescriptorService, instantiationService, openerService, themeService, hoverService);
-
-		this.viewModel = changesViewService.viewModel;
 
 		// Context keys
 		this.isMergeBaseBranchProtectedContextKey = ActiveSessionContextKeys.IsMergeBaseBranchProtected.bindTo(this.scopedContextKeyService);
@@ -424,19 +419,19 @@ export class ChangesViewPane extends ViewPane {
 
 		// Version mode
 		this._register(bindContextKey(ChangesContextKeys.VersionMode, this.scopedContextKeyService, reader => {
-			return this.viewModel.activeSessionChangesetObs.read(reader)?.id ?? '';
+			return this.changesViewService.activeSessionChangesetObs.read(reader)?.id ?? '';
 		}));
 
 		// View mode
 		this._register(bindContextKey(ChangesContextKeys.ViewMode, this.scopedContextKeyService, reader => {
-			return this.viewModel.viewModeObs.read(reader);
+			return this.changesViewService.viewModeObs.read(reader);
 		}));
 
 		// Set chatSessionType on the view's context key service so ViewTitle menu items
 		// can use it in their `when` clauses. Update reactively when the active session
 		// changes.
 		this._register(bindContextKey(ChatContextKeys.agentSessionType, this.scopedContextKeyService, reader => {
-			return this.viewModel.activeSessionTypeObs.read(reader) ?? '';
+			return this.changesViewService.activeSessionTypeObs.read(reader) ?? '';
 		}));
 
 		// Git operation in progress set in the global context key service by the extension
@@ -446,7 +441,7 @@ export class ChangesViewPane extends ViewPane {
 
 		// Git operation in progress set in the session state
 		const hasGitOperationInProgressStateObs = derived(reader => {
-			const activeSessionState = this.viewModel.activeSessionStateObs.read(reader);
+			const activeSessionState = this.changesViewService.activeSessionStateObs.read(reader);
 			return activeSessionState?.hasGitOperationInProgress === true;
 		});
 
@@ -503,7 +498,7 @@ export class ChangesViewPane extends ViewPane {
 			menuOptions: { shouldForwardArgs: true },
 			actionViewItemProvider: (action) => {
 				if (action.id === 'chatEditing.versionsPicker' && action instanceof MenuItemAction) {
-					return this.scopedInstantiationService.createInstance(ChangesPickerActionItem, action, this.viewModel);
+					return this.scopedInstantiationService.createInstance(ChangesPickerActionItem, action);
 				}
 				return undefined;
 			},
@@ -515,7 +510,7 @@ export class ChangesViewPane extends ViewPane {
 			menuOptions: { shouldForwardArgs: true },
 			actionViewItemProvider: (action, options) => {
 				if (action.id === ChangesDiffStatsAction.ID && action instanceof MenuItemAction) {
-					return this.scopedInstantiationService.createInstance(ChangesDiffStatsActionItem, action, this.viewModel, options);
+					return this.scopedInstantiationService.createInstance(ChangesDiffStatsActionItem, action, options);
 				}
 				return undefined;
 			},
@@ -636,7 +631,7 @@ export class ChangesViewPane extends ViewPane {
 	}
 
 	override getActionsContext(): URI | undefined {
-		return this.viewModel.activeSessionResourceObs.get();
+		return this.changesViewService.activeSessionResourceObs.get();
 	}
 
 	private onVisible(): void {
@@ -644,13 +639,13 @@ export class ChangesViewPane extends ViewPane {
 
 		// Title actions
 		this.renderDisposables.add(autorun(reader => {
-			this.viewModel.activeSessionResourceObs.read(reader);
+			this.changesViewService.activeSessionResourceObs.read(reader);
 			this.updateActions();
 		}));
 
 		// Loading
 		this.renderDisposables.add(autorun(reader => {
-			const isLoading = this.viewModel.activeSessionIsLoadingObs.read(reader);
+			const isLoading = this.changesViewService.activeSessionIsLoadingObs.read(reader);
 			if (isLoading) {
 				this.changesProgressBar.infinite().show(200);
 			} else {
@@ -660,13 +655,13 @@ export class ChangesViewPane extends ViewPane {
 
 		// Changes
 		const changesObs = derived(reader => {
-			const changes = this.viewModel.activeSessionChangesObs.read(reader);
+			const changes = this.changesViewService.activeSessionChangesObs.read(reader);
 			return toIChangesFileItem(changes);
 		});
 
 		// Changes statistics
 		const topLevelStats = derivedObservableWithCache<{ files: number; added: number; removed: number } | undefined>(this, (reader, lastValue) => {
-			const isLoading = this.viewModel.activeSessionIsLoadingObs.read(reader);
+			const isLoading = this.changesViewService.activeSessionIsLoadingObs.read(reader);
 			if (isLoading) {
 				return lastValue;
 			}
@@ -699,8 +694,8 @@ export class ChangesViewPane extends ViewPane {
 				const isAgentHostSession = isAgentHostSessionObs.read(reader);
 
 				const widget = isAgentHostSession
-					? this.scopedInstantiationService.createInstance(ChangesWorkbenchButtonBarWidget, this.actionsContainer!, this.viewModel)
-					: this.scopedInstantiationService.createInstance(ChangesMenuWorkbenchButtonBarWidget, this.actionsContainer!, this.viewModel, this.hasGitOperationInProgressObs);
+					? this.scopedInstantiationService.createInstance(ChangesWorkbenchButtonBarWidget, this.actionsContainer!)
+					: this.scopedInstantiationService.createInstance(ChangesMenuWorkbenchButtonBarWidget, this.actionsContainer!, this.hasGitOperationInProgressObs);
 				reader.store.add(widget);
 			}));
 		}
@@ -712,7 +707,7 @@ export class ChangesViewPane extends ViewPane {
 
 		// Update visibility based on entries
 		this.renderDisposables.add(autorun(reader => {
-			if (this.viewModel.activeSessionIsLoadingObs.read(reader)) {
+			if (this.changesViewService.activeSessionIsLoadingObs.read(reader)) {
 				return;
 			}
 
@@ -723,7 +718,7 @@ export class ChangesViewPane extends ViewPane {
 				dom.setVisibility(!isUntitled, this.actionsContainer);
 			}
 
-			const hasGitRepository = this.viewModel.activeSessionHasGitRepositoryObs.read(reader);
+			const hasGitRepository = this.changesViewService.activeSessionHasGitRepositoryObs.read(reader);
 
 			const stats = topLevelStats.read(reader);
 			const hasEntries = stats !== undefined && stats.files > 0;
@@ -795,11 +790,11 @@ export class ChangesViewPane extends ViewPane {
 		// Update tree data with combined entries
 		this.renderDisposables.add(autorun(reader => {
 			const changes = changesObs.read(reader);
-			const viewMode = this.viewModel.viewModeObs.read(reader);
-			const isLoading = this.viewModel.activeSessionIsLoadingObs.read(reader);
+			const viewMode = this.changesViewService.viewModeObs.read(reader);
+			const isLoading = this.changesViewService.activeSessionIsLoadingObs.read(reader);
 			// Read session state so this autorun re-runs when git state (e.g. branch name)
 			// arrives asynchronously, since the tree root label depends on it.
-			this.viewModel.activeSessionStateObs.read(reader);
+			this.changesViewService.activeSessionStateObs.read(reader);
 
 			if (!this.tree || isLoading) {
 				return;
@@ -841,7 +836,7 @@ export class ChangesViewPane extends ViewPane {
 
 		// Bulk update the context keys
 		this.renderDisposables.add(autorun(reader => {
-			const state = this.viewModel.activeSessionStateObs.read(reader);
+			const state = this.changesViewService.activeSessionStateObs.read(reader);
 			if (!state || state.hasGitOperationInProgress) {
 				return;
 			}
@@ -925,7 +920,7 @@ export class ChangesViewPane extends ViewPane {
 			name = `${segments.slice(0, 2).join('/')} (${decodeURIComponent(segments[2])})`;
 		} else {
 			// Local session
-			const branchName = this.viewModel.activeSessionStateObs.get()?.branchName;
+			const branchName = this.changesViewService.activeSessionStateObs.get()?.branchName;
 			name = branchName
 				? `${basename(folder.workingDirectory)} (${branchName})`
 				: basename(folder.workingDirectory);
@@ -942,7 +937,7 @@ export class ChangesViewPane extends ViewPane {
 	}
 
 	private getSessionDiscardRef(): string {
-		const changeset = this.viewModel.activeSessionChangesetObs.get();
+		const changeset = this.changesViewService.activeSessionChangesetObs.get();
 		return changeset?.originalCheckpointRef.get() ?? '';
 	}
 
@@ -972,7 +967,7 @@ export class ChangesViewPane extends ViewPane {
 
 		container.classList.add('changes-file-list');
 
-		const viewMode = this.viewModel.viewModeObs.get();
+		const viewMode = this.changesViewService.viewModeObs.get();
 		container.classList.toggle('list-mode', viewMode === ChangesViewMode.List);
 
 		// "Changes" header
@@ -1052,7 +1047,7 @@ export class ChangesViewPane extends ViewPane {
 
 		const resourceLabels = disposables.add(this.instantiationService.createInstance(ResourceLabels, { onDidChangeVisibility }));
 		const actionRunner = disposables.add(new ChangesViewActionRunner(
-			() => this.viewModel.activeSessionResourceObs.get(),
+			() => this.changesViewService.activeSessionResourceObs.get(),
 			() => this.getSessionDiscardRef(),
 			getSelection ?? (() => this.getTreeSelection()),
 		));
@@ -1061,7 +1056,7 @@ export class ChangesViewPane extends ViewPane {
 			'ChangesViewTree',
 			container,
 			new ChangesTreeDelegate(),
-			[this.instantiationService.createInstance(ChangesTreeRenderer, this.viewModel, resourceLabels, actionRunner,
+			[this.instantiationService.createInstance(ChangesTreeRenderer, resourceLabels, actionRunner,
 				() => {
 					// Pass in the tree root to be used to compute the label description
 					const activeSession = this.sessionsService.activeSession.get();
@@ -1101,11 +1096,11 @@ export class ChangesViewPane extends ViewPane {
 				identityProvider: {
 					getId: (element: ChangesTreeElement) => element.uri.toString()
 				},
-				indent: this.viewModel.viewModeObs.get() === ChangesViewMode.List ? 0 : 8,
+				indent: this.changesViewService.viewModeObs.get() === ChangesViewMode.List ? 0 : 8,
 				compressionEnabled: true,
-				sorter: new ChangesTreeSorter(() => this.viewModel.viewModeObs.get()),
+				sorter: new ChangesTreeSorter(() => this.changesViewService.viewModeObs.get()),
 				twistieAdditionalCssClass: (e: unknown) => {
-					return this.viewModel.viewModeObs.get() === ChangesViewMode.List
+					return this.changesViewService.viewModeObs.get() === ChangesViewMode.List
 						? 'force-no-twistie'
 						: undefined;
 				},
@@ -1114,7 +1109,7 @@ export class ChangesViewPane extends ViewPane {
 	}
 
 	async openChanges(resource?: URI): Promise<void> {
-		const items = this.viewModel.activeSessionChangesObs.get();
+		const items = this.changesViewService.activeSessionChangesObs.get();
 		if (items.length === 0) {
 			return;
 		}
@@ -1211,8 +1206,8 @@ export class ChangesViewPane extends ViewPane {
 	}
 
 	private async _openMultiFileDiffEditor(reveal?: URI): Promise<void> {
-		const sessionResource = this.viewModel.activeSessionResourceObs.get();
-		const changes = this.viewModel.activeSessionChangesObs.get();
+		const sessionResource = this.changesViewService.activeSessionResourceObs.get();
+		const changes = this.changesViewService.activeSessionChangesObs.get();
 
 		if (!sessionResource || changes.length === 0) {
 			return;
@@ -1376,7 +1371,7 @@ class SetChangesListViewModeAction extends ViewAction<ChangesViewPane> {
 
 	async runInView(accessor: ServicesAccessor, view: ChangesViewPane): Promise<void> {
 		logChangesViewViewModeChange(accessor.get(ITelemetryService), ChangesViewMode.List);
-		view.viewModel.setViewMode(ChangesViewMode.List);
+		accessor.get(IChangesViewService).setViewMode(ChangesViewMode.List);
 	}
 }
 
@@ -1399,7 +1394,7 @@ class SetChangesTreeViewModeAction extends ViewAction<ChangesViewPane> {
 
 	async runInView(accessor: ServicesAccessor, view: ChangesViewPane): Promise<void> {
 		logChangesViewViewModeChange(accessor.get(ITelemetryService), ChangesViewMode.Tree);
-		view.viewModel.setViewMode(ChangesViewMode.Tree);
+		accessor.get(IChangesViewService).setViewMode(ChangesViewMode.Tree);
 	}
 }
 
@@ -1434,16 +1429,16 @@ registerAction2(VersionsPickerAction);
 class ChangesPickerActionItem extends ActionWidgetDropdownActionViewItem {
 	constructor(
 		action: MenuItemAction,
-		private readonly viewModel: ChangesViewModel,
 		@IActionWidgetService actionWidgetService: IActionWidgetService,
 		@IKeybindingService keybindingService: IKeybindingService,
 		@IContextKeyService contextKeyService: IContextKeyService,
+		@IChangesViewService private readonly changesViewService: IChangesViewService,
 		@ITelemetryService private readonly telemetryService: ITelemetryService,
 	) {
 		const actionProvider: IActionWidgetDropdownActionProvider = {
 			getActions: () => {
-				const changesets = viewModel.activeSessionChangesetsObs.get() ?? [];
-				const selectedChangeset = viewModel.activeSessionChangesetObs.get();
+				const changesets = changesViewService.activeSessionChangesetsObs.get() ?? [];
+				const selectedChangeset = changesViewService.activeSessionChangesetObs.get();
 
 				return changesets.map(changeset => ({
 					...action,
@@ -1458,7 +1453,7 @@ class ChangesPickerActionItem extends ActionWidgetDropdownActionViewItem {
 					},
 					enabled: changeset.isEnabled.get(),
 					run: async () => {
-						viewModel.setChangesetId(changeset.id);
+						changesViewService.setChangesetId(changeset.id);
 						logChangesViewVersionModeChange(this.telemetryService, changeset.id);
 					}
 				} satisfies IActionWidgetDropdownAction));
@@ -1468,7 +1463,7 @@ class ChangesPickerActionItem extends ActionWidgetDropdownActionViewItem {
 		super(action, { actionProvider, listOptions: { detailItemHeight: 44 } }, actionWidgetService, keybindingService, contextKeyService, telemetryService);
 
 		this._register(autorun(reader => {
-			viewModel.activeSessionChangesetObs.read(reader);
+			changesViewService.activeSessionChangesetObs.read(reader);
 
 			if (this.element) {
 				this.renderLabel(this.element);
@@ -1477,7 +1472,7 @@ class ChangesPickerActionItem extends ActionWidgetDropdownActionViewItem {
 	}
 
 	protected override renderLabel(element: HTMLElement): IDisposable | null {
-		const changeset = this.viewModel.activeSessionChangesetObs.get();
+		const changeset = this.changesViewService.activeSessionChangesetObs.get();
 		if (!changeset) {
 			return null;
 		}
@@ -1544,15 +1539,15 @@ class ChangesDiffStatsActionItem extends ActionViewItem {
 
 	constructor(
 		action: MenuItemAction,
-		viewModel: ChangesViewModel,
 		options: IActionViewItemOptions,
+		@IChangesViewService changesViewService: IChangesViewService
 	) {
 		super(null, action, { ...options, icon: false, label: true });
 
 		const diffStatsRawObs = derivedObservableWithCache<{ files: number; insertions: number; deletions: number } | undefined>(this,
 			(reader, lastValue) => {
-				const entries = viewModel.activeSessionChangesObs.read(reader);
-				const isLoading = viewModel.activeSessionIsLoadingObs.read(reader);
+				const entries = changesViewService.activeSessionChangesObs.read(reader);
+				const isLoading = changesViewService.activeSessionIsLoadingObs.read(reader);
 
 				if (isLoading) {
 					return lastValue;
